+++ apps/control/src/config/control/fc002Config.ts
